fix: breakdown of big file
